requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
dash==0.43.0
dash-core-components==0.48.0
dash-html-components==0.16.0
dash-renderer==0.24.0
dash-table==3.7.0
decorator==4.4.0
Django==2.2.1
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Compress==1.4.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
image==1.5.27
ipython-genutils==0.2.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mock==3.0.5
nbformat==4.4.0
numpy==1.16.3
Pillow==6.0.0
plotly==3.9.0
protobuf==3.7.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
retrying==1.3.3
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.24.3
Werkzeug==0.15.2
wrapt==1.11.1
